# backoffice/frontend/openapi.yaml
openapi: 3.0.3
info:
  title: Demo API
  description: Demo API for React application
  version: 1.0.0
servers:
  - url: http://localhost:8080/api
    description: Local server
paths:
  /users:
    get:
      operationId: getUsers
      summary: Get all users
      tags:
        - Users
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsersList'
    post:
      operationId: createUser
      summary: Create a new user
      tags:
        - Users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserInput'
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
...
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: "John Doe"
        email:
          type: string
          format: email
          example: "john@example.com"
        role:
          type: string
          enum: [admin, user, manager]
          example: "user"
        active:
          type: boolean
          example: true
        createdAt:
          type: string
          format: date-time
    UserInput:
      type: object
      required:
        - name
        - email
        - role
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        role:
          type: string
          enum: [admin, user, manager]
        active:
          type: boolean
    UsersList:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/User'
        totalElements:
          type: integer
        totalPages:
          type: integer
...